% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bindCensusDataToMap.R
\name{bindCensusDataToMap}
\alias{bindCensusDataToMap}
\title{Binds Census data to S4 dataframe}
\usage{
bindCensusDataToMap(df, spdf)
}
\arguments{
\item{df}{a census dataframe}

\item{spdf}{an S4 dataframe}
}
\value{
A spatial polygons dataframe with extra columns in @data
}
\description{
Binds Census data to a Spatial Polygons Dataframe from tigris package using GEOID as unique identifier.
}

