% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createCensusURL.R
\name{createCensusURL}
\alias{createCensusURL}
\title{Create and API call URL}
\usage{
createCensusURL(year, vars, tract = "all", county = "all", state = "all",
  level = "county", key)
}
\arguments{
\item{year}{must be 1990, 2000 of 2010}

\item{vars}{a string or vector of strings specifying which variables to download. Must be the exact variable codes.}

\item{tract}{a string or vector of strings specifying for which tracts to download data. Must be exact FIPS code.}

\item{county}{a string or vector of strings specifying for which counties to download data. Must be exact FIPS code}

\item{state}{FIPS code or vector of FIPS codes specifying for which states to download data.}

\item{level}{can be either 'county' or 'tract', specifying at which geographic level to download data}

\item{key}{your personal API key}
}
\value{
a string
}
\description{
creates the URL with which to query the Census API for the specified variables and geographic levels. Only functional for the decennial censuses.
}

