% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/createCleanCensusDF.R
\name{createCleanCensusDF}
\alias{createCleanCensusDF}
\title{Tidies and downloads Census data}
\usage{
createCleanCensusDF(url, varNames = "", tracts = FALSE)
}
\arguments{
\item{url}{a URL generated from createCensusURL}

\item{varNames}{a string or vector of strings}

\item{tracts}{a logical corresponding to whether data is being downloaded at the tract level}
}
\value{
a dataframe
}
\description{
create a clean dataset from the Census API using a URL generated in a createCensusURl.
}

